Preparing for upcoming release. Pushing latest code to staging environment.

+++ src/main/java/gov/epa/bencloud/server/tasks/TaskUtil.java
@@ -17,7 +17,7 @@ public class TaskUtil {
 	private static final Logger log = LoggerFactory.getLogger(TaskUtil.class);
 
 	
-	public static void deleteHifResults(String uuid) {
+	public static void deleteHifResults(String uuid, Boolean deleteTask) {
 		
 		try {
 
@@ -29,7 +29,7 @@ public class TaskUtil {
 						.where(HIF_RESULT_DATASET.TASK_UUID.eq(uuid))
 						.fetch();
 				
-				if (hifResultDatasets.size() == 0) {
+				if (hifResultDatasets.size() == 0 && deleteTask) {
 					DSL.using(ctx).deleteFrom(TASK_COMPLETE)
 					.where(TASK_COMPLETE.TASK_UUID.eq(uuid))
 					.execute();
@@ -39,10 +39,10 @@ public class TaskUtil {
 
 					Record hifResultDataset = hifResultDatasets.get(0);
 
-					Result<Record1<Integer>> hifResults = 
-							DSL.using(ctx).select(HIF_RESULT.HIF_RESULT_DATASET_ID).from(HIF_RESULT)
-					.where(HIF_RESULT.HIF_RESULT_DATASET_ID.eq(hifResultDataset.get(HIF_RESULT_DATASET.ID)))
-					.fetch();
+//					Result<Record1<Integer>> hifResults = 
+//							DSL.using(ctx).select(HIF_RESULT.HIF_RESULT_DATASET_ID).from(HIF_RESULT)
+//					.where(HIF_RESULT.HIF_RESULT_DATASET_ID.eq(hifResultDataset.get(HIF_RESULT_DATASET.ID)))
+//					.fetch();
 
 					DSL.using(ctx).deleteFrom(HIF_RESULT)
 					.where(HIF_RESULT.HIF_RESULT_DATASET_ID.eq(hifResultDataset.get(HIF_RESULT_DATASET.ID)))
@@ -55,10 +55,14 @@ public class TaskUtil {
 					DSL.using(ctx).deleteFrom(HIF_RESULT_FUNCTION_CONFIG)
 					.where(HIF_RESULT_FUNCTION_CONFIG.HIF_RESULT_DATASET_ID.eq(hifResultDataset.get(HIF_RESULT_DATASET.ID)))
 					.execute();
+					
+					if(deleteTask) {
+						DSL.using(ctx).deleteFrom(TASK_COMPLETE)
+						.where(TASK_COMPLETE.TASK_UUID.eq(uuid))
+						.execute();
+					}
 
-					DSL.using(ctx).deleteFrom(TASK_COMPLETE)
-					.where(TASK_COMPLETE.TASK_UUID.eq(uuid))
-					.execute();
+					
 				}
 			});
 		} catch (Exception e) {
@@ -69,7 +73,7 @@ public class TaskUtil {
 
 	}
 
-	public static void deleteValuationResults(String uuid) {
+	public static void deleteValuationResults(String uuid, Boolean deleteTask) {
 		
 		try {
 
@@ -81,7 +85,7 @@ public class TaskUtil {
 						.where(VALUATION_RESULT_DATASET.TASK_UUID.eq(uuid))
 						.fetch();
 
-				if (valuationResultDatasets.size() == 0) {
+				if (valuationResultDatasets.size() == 0 && deleteTask) {
 					DSL.using(ctx).deleteFrom(TASK_COMPLETE)
 					.where(TASK_COMPLETE.TASK_UUID.eq(uuid))
 					.execute();
@@ -91,10 +95,10 @@ public class TaskUtil {
 
 					Record valuationResultDataset = valuationResultDatasets.get(0);
 
-					Result<Record1<Integer>> ValuationResults = 
-							DSL.using(ctx).select(VALUATION_RESULT.VALUATION_RESULT_DATASET_ID).from(VALUATION_RESULT)
-					.where(VALUATION_RESULT.VALUATION_RESULT_DATASET_ID.eq(valuationResultDataset.get(VALUATION_RESULT_DATASET.ID)))
-					.fetch();
+//					Result<Record1<Integer>> ValuationResults = 
+//							DSL.using(ctx).select(VALUATION_RESULT.VALUATION_RESULT_DATASET_ID).from(VALUATION_RESULT)
+//					.where(VALUATION_RESULT.VALUATION_RESULT_DATASET_ID.eq(valuationResultDataset.get(VALUATION_RESULT_DATASET.ID)))
+//					.fetch();
 
 					DSL.using(ctx).deleteFrom(VALUATION_RESULT)
 					.where(VALUATION_RESULT.VALUATION_RESULT_DATASET_ID.eq(valuationResultDataset.get(VALUATION_RESULT_DATASET.ID)))
@@ -107,10 +111,14 @@ public class TaskUtil {
 					DSL.using(ctx).deleteFrom(VALUATION_RESULT_FUNCTION_CONFIG)
 					.where(VALUATION_RESULT_FUNCTION_CONFIG.VALUATION_RESULT_DATASET_ID.eq(valuationResultDataset.get(VALUATION_RESULT_DATASET.ID)))
 					.execute();
+					
+					if(deleteTask) {
+						DSL.using(ctx).deleteFrom(TASK_COMPLETE)
+						.where(TASK_COMPLETE.TASK_UUID.eq(uuid))
+						.execute();
+					}
 
-					DSL.using(ctx).deleteFrom(TASK_COMPLETE)
-					.where(TASK_COMPLETE.TASK_UUID.eq(uuid))
-					.execute();
+					
 				}
 			});
 		} catch (Exception e) {
@@ -121,4 +129,54 @@ public class TaskUtil {
 
 	}
 
+	public static void deleteExposureResults(String uuid, Boolean deleteTask) {
+		
+		try {
+
+			DSL.using(JooqUtil.getJooqConfiguration())
+					.transaction(ctx -> {
+
+				Result<Record1<Integer>> exposureResultDatasets = 
+						DSL.using(ctx).select(EXPOSURE_RESULT_DATASET.ID).from(EXPOSURE_RESULT_DATASET)
+						.where(EXPOSURE_RESULT_DATASET.TASK_UUID.eq(uuid))
+						.fetch();
+
+				if (exposureResultDatasets.size() == 0 && deleteTask) {
+					DSL.using(ctx).deleteFrom(TASK_COMPLETE)
+					.where(TASK_COMPLETE.TASK_UUID.eq(uuid))
+					.execute();
+				} else if (exposureResultDatasets.size() > 1) {
+					log.info("recieved more than 1 Exposure Result Dataset record");
+				} else {
+
+					Record exposureResultDataset = exposureResultDatasets.get(0);
+
+					DSL.using(ctx).deleteFrom(EXPOSURE_RESULT)
+					.where(EXPOSURE_RESULT.EXPOSURE_RESULT_DATASET_ID.eq(exposureResultDataset.get(EXPOSURE_RESULT_DATASET.ID)))
+					.execute();
+					
+					DSL.using(ctx).deleteFrom(EXPOSURE_RESULT_DATASET)
+					.where(EXPOSURE_RESULT_DATASET.ID.eq(exposureResultDataset.get(EXPOSURE_RESULT_DATASET.ID)))
+					.execute();
+
+					DSL.using(ctx).deleteFrom(EXPOSURE_RESULT_FUNCTION_CONFIG)
+					.where(EXPOSURE_RESULT_FUNCTION_CONFIG.EXPOSURE_RESULT_DATASET_ID.eq(exposureResultDataset.get(EXPOSURE_RESULT_DATASET.ID)))
+					.execute();
+					
+					if(deleteTask) {
+						DSL.using(ctx).deleteFrom(TASK_COMPLETE)
+						.where(TASK_COMPLETE.TASK_UUID.eq(uuid))
+						.execute();
+					}
+
+					
+				}
+			});
+		} catch (Exception e) {
+			log.error("Error deleting exposure results", e);
+		} finally {
+
+		}
+
+	}
 }
